+++ Firmware/planner.h
@@ -187,8 +187,8 @@ extern float* max_feedrate;
 
 
 // Use M201 to override by software
-extern unsigned long* max_acceleration_mm_per_s2;
-extern unsigned long max_acceleration_steps_per_s2[NUM_AXIS];
+extern uint32_t* max_acceleration_mm_per_s2; 
+extern uint32_t max_acceleration_steps_per_s2[NUM_AXIS];
 
 extern long position[NUM_AXIS];
